netlink: connector: Add idx and val identifiers for netfilter D-Bus
The D-bus IPC system implements a transport that uses AF_BUS sockets to
send D-Bus messages to the peers. This allows decouple the routing logic
from the daemon and move it to the kernel which has the advantage of
reducing the number of context switches and the messages copied to
user-space.

A D-Bus protocol aware netfilter module decide which peer can recive a
given message based on a set of D-Bus match rules. These match rules
are set from user-space using the netlink connector API.

Based on a previous patch by Alban Crequy.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>