mach-shmobile: Emma Mobile EV2 DT support V3
This is EMEV2 DT support V3. The support is limited to
whatever devices that are complied in the kernel. At this
point we have UARTs handled by "em-uart" and a timer
handled by "em-sti". Clocks and SMP are not supported.

Signed-off-by: Magnus Damm <damm@opensource.se>
Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>
(cherry picked from commit 3d5de27174955702bc874302ba9e72d71d5acd58)

Signed-off-by: Tetsuyuki Kobayashi <koba@kmckk.co.jp>