drm/rcar-du: Support configurable color keying for planes
The color key is expressed as an RGB888 triplet stored in a 32-bit
integer in XRGB8888 format. Bit 24 is used as a flag to disable (0) or
enable color keying (1). The hardware supports source color keying only.

Create a new colorkey property and instantiate it for each plane with an
initial value of 0.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
(cherry picked from commit 1139f309ec6e1123b3500bfef4cbb920a6196112)

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>